fix conflict after merge

+++ domain/search/src/main/java/com/domain/search/useCases/FilterMediaByRatingUseCase.kt
@@ -3,7 +3,7 @@ package com.domain.search.useCases
 import com.domain.search.model.Media
 
 class FilterMediaByRatingUseCase {
-    operator fun invoke(rating: Double, mediaList: List<Media>): List<Media> {
+    operator fun invoke(rating: Float, mediaList: List<Media>): List<Media> {
         return mediaList.filter { media -> media.rating >= rating }
     }
 }
